refactor: seat, seatGrade 분리
